Move user profile management to a separate app server.user_profile
