+++ openquake/server/settings.py
@@ -37,12 +37,7 @@ WEBUI_USER = 'openquake'
 
 TEST = 'test' in sys.argv
 
-# NOTE: the UserProfile class makes it more complicated to add
-# the apps django.contrib.auth and django.contrib.contenttypes conditionally
-# to the fact that the authentication is required. However, it is not a problem
-# to exclude AuthenticationMiddleware and LoginRequiredMiddleware in PUBLIC mode
-INSTALLED_APPS = ('openquake.server.db', 'django.contrib.auth',
-                  'django.contrib.contenttypes')
+INSTALLED_APPS = ('openquake.server.db',)
 
 OQSERVER_ROOT = os.path.dirname(__file__)
 
@@ -289,7 +284,8 @@ APPLICATION_MODE = os.environ.get('OQ_APPLICATION_MODE', APPLICATION_MODE)
 
 if APPLICATION_MODE not in ('PUBLIC',):
     # add installed_apps for cookie-consent
-    for app in ('cookie_consent',):
+    for app in ('django.contrib.auth', 'django.contrib.contenttypes',
+                'openquake.server.user_profile', 'cookie_consent',):
         if app not in INSTALLED_APPS:
             INSTALLED_APPS += (app,)
 
@@ -364,7 +360,9 @@ if LOCKDOWN:
         'openquake.server.middleware.LoginRequiredMiddleware',
     )
 
-    for app in (
+    for app in ('django.contrib.auth',
+                'django.contrib.contenttypes',
+                'openquake.server.user_profile',
                 'django.contrib.messages',
                 'django.contrib.sessions',
                 'django.contrib.admin',
